+++ src/io.rs
@@ -1,6 +1,6 @@
 use std::io;
 
-use crate::{limits, util::u8_from_hex_digit};
+use crate::util::u8_from_hex_digit;
 
 fn decode_byte(pair: &[u8; 2]) -> Result<u8, crate::Error> {
     Ok(u8_from_hex_digit(pair[0])? << 4 | u8_from_hex_digit(pair[1])?)
@@ -10,7 +10,7 @@ pub(crate) trait MyReader {
     type Error: From<crate::Error>;
 
     fn read_bytes<const N: usize>(&mut self) -> Result<[u8; N], Self::Error>;
-    fn read_vec(&mut self, len: u64) -> Result<Vec<u8>, Self::Error>;
+    fn read_vec(&mut self, len: u64, oom_mitigation: usize) -> Result<Vec<u8>, Self::Error>;
 }
 
 #[repr(transparent)]
@@ -25,9 +25,11 @@ impl<'a> MyReader for SliceReader<'a> {
         Ok(*bytes)
     }
 
-    fn read_vec(&mut self, len: u64) -> Result<Vec<u8>, Self::Error> {
+    fn read_vec(&mut self, len: u64, _oom_mitigation: usize) -> Result<Vec<u8>, Self::Error> {
         let len = usize::try_from(len).or(Err(crate::Error::LengthTooLarge))?;
         let slice = self.0.split_off(..len).ok_or(crate::Error::UnexpectedEof)?;
+
+        // No OOM mitigation necessary: split_off() was successful
         Ok(slice.to_vec())
     }
 }
@@ -49,12 +51,14 @@ impl<'a> MyReader for HexSliceReader<'a> {
         Ok(buf)
     }
 
-    fn read_vec(&mut self, len: u64) -> Result<Vec<u8>, Self::Error> {
+    fn read_vec(&mut self, len: u64, _oom_mitigation: usize) -> Result<Vec<u8>, Self::Error> {
         let len = usize::try_from(len).or(Err(crate::Error::LengthTooLarge))?;
         let hex_len = len.checked_mul(2).ok_or(crate::Error::LengthTooLarge)?;
         let hex = self.0.split_off(..hex_len).ok_or(crate::Error::UnexpectedEof)?;
 
+        // No OOM mitigation necessary: split_off() was successful
         let mut vec = Vec::with_capacity(len);
+
         for pair in hex.as_chunks::<2>().0 {
             vec.push(decode_byte(pair)?);
         }
@@ -73,11 +77,11 @@ impl<R: io::Read> MyReader for R {
         Ok(buf)
     }
 
-    fn read_vec(&mut self, len: u64) -> Result<Vec<u8>, Self::Error> {
+    fn read_vec(&mut self, len: u64, oom_mitigation: usize) -> Result<Vec<u8>, Self::Error> {
         use io::Read;
 
         let len_usize = usize::try_from(len).or(Err(crate::Error::LengthTooLarge))?;
-        let mut buf = Vec::with_capacity(len_usize.min(limits::OOM_MITIGATION)); // Mitigate OOM
+        let mut buf = Vec::with_capacity(len_usize.min(oom_mitigation));
         let bytes_read = self.take(len).read_to_end(&mut buf)?;
 
         if bytes_read == len_usize {
@@ -106,9 +110,9 @@ impl<R: io::Read> MyReader for HexReader<R> {
         Ok(buf)
     }
 
-    fn read_vec(&mut self, len: u64) -> Result<Vec<u8>, Self::Error> {
+    fn read_vec(&mut self, len: u64, oom_mitigation: usize) -> Result<Vec<u8>, Self::Error> {
         let len_usize = usize::try_from(len).or(Err(crate::Error::LengthTooLarge))?;
-        let mut vec = Vec::with_capacity(len_usize.min(limits::OOM_MITIGATION));
+        let mut vec = Vec::with_capacity(len_usize.min(oom_mitigation));
 
         let mut buf = [0_u8; 1024];
         let mut remaining = len_usize;